created deleting option for availble timeslots using calendar section and devided calendar component into small components
